add cache command
